[#34] refactor: action 파일 분리
- reducer도 날짜 관련, 인원 관련으로 2개가 있기 때문에
  action도 2개로 분리함
