+++ FE/src/actions/datePickerAction.js
@@ -1,0 +1,6 @@
+export const dateChange = (payload) => {
+  return {
+    type: "dateChange",
+    payload,
+  };
+};
