+++ FE/src/actions/guestCountAction.js
@@ -20,10 +20,3 @@ export const deleteCount = (payload) => {
     payload,
   };
 };
-
-export const dateChange = (payload) => {
-  return {
-    type: "dateChange",
-    payload,
-  };
-};
